Delete some improperly imported files
