+++ server/node-server/app.js
@@ -5,19 +5,26 @@ import cors from "cors";
 import alertRoute from "./routes/alert.js";
 import homepage from './routes/homepage.js';
 
-const port = 8000;
+const port = 8080;
 const app = express();
 dotenv.config();
 
+const corsOptions = {
+  origin: 'https://phishjail-api-gateway-82gpumlr.ew.gateway.dev',
+  methods: 'GET,HEAD,PUT,PATCH,POST,DELETE',
+  optionsSuccessStatus: 201,
+  allowedHeaders: 'Content-Type,Authorization',
+};
+
 //middlewares
-app.use(cors())
+app.use(cors(corsOptions))
 app.use(cookieParser())
 app.use(express.json());
 
 app.use(homepage, alertRoute)
 
 app.get('/status', (req, res) => {
-  res.status(200).json({ status: "healthy" });
+  res.status(201).json({ status: "healthy" });
 });
 
 app.use((err, req, res) => {
