--- server/node-server/node-server/.env
@@ -1,1 +1,0 @@
-KEY=''
