--- server/node-server/node-server/controllers/homepage.js
@@ -1,8 +1,0 @@
-export const homepage = async (req, res, next) => {
-  try {
-    res.send('<h1>Received a credential reuse alert home!</h1>');
-  } catch (err) {
-    next(err);
-  }
-  next()
-};
