--- server/node-server/node-server/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "server",
-  "version": "1.0.0",
-  "description": "",
-  "main": "app.js",
-  "type": "module",
-  "scripts": {
-    "test": "jest",
-    "start": "nodemon app.js"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "body-parser": "^1.20.2",
-    "cookie-parser": "^1.4.6",
-    "cors": "^2.8.5",
-    "dotenv": "^16.3.1",
-    "express": "^4.18.2",
-    "nodemon": "^3.0.1",
-    "winston": "^3.10.0"
-  },
-  "nodemonConfig": {
-    "ignore": [
-      "**/test/**",
-      "**/docs/**"
-    ],
-    "delay": 2500
-  },
-  "devDependencies": {
-    "jest": "^29.6.4",
-    "supertest": "^6.3.3"
-  }
-}
