--- server/node-server/node-server/routes/alert.js
@@ -1,8 +1,0 @@
-import express from "express";
-import { alert } from "../controllers/alert.js";
-
-const alertRoute = express.Router();
-
-alertRoute.post("/alert", alert)
-
-export default alertRoute
