--- server/node-server/node-server/routes/homepage.js
@@ -1,8 +1,0 @@
-import express from "express";
-import { homepage } from "../controllers/homepage.js";
-
-const home = express.Router();
-
-home.get("/", homepage)
-
-export default home
