+++ server/node-server/tests/app.test.js
@@ -1,12 +1,16 @@
-import { describe } from 'jest';
-import { alert } from "../controllers/alert.js";
+const axios = require('axios')
 
+const baseURL = 'http://localhost:8080';
+const headers = {
+  'Content-Type': 'application/json',
+  'x-api-key': process.env.API_KEY
+}
 describe('API Server Test', () => {
   it('POST /alert with correct PSK should return alert success', async () => {
     const alertData = {
       allAssociatedUsernames: "bob",
       alertUrl: "https://www.example.com",
-      psk: process.env.KEY,
+      key: process.env.API_KEY,
       referrer: "https://www.google.com",
       alertType: "reuse",
       suspectedUsername: "testuser",
@@ -15,9 +19,9 @@ describe('API Server Test', () => {
       clientId: "foo"
     };
 
-    const response = await request(alert).post('/alert').send(alertData);
-    expect(response.statusCode).toBe(200);
-    expect(response.body.status).toBe('alert success');
+    const response = await axios.post(`${baseURL}/alert`, alertData, { headers });
+    expect(response.status).toBe(200);
+    expect(response.data.status).toBe('Alert success');
   });
 
 
@@ -25,7 +29,7 @@ describe('API Server Test', () => {
     const alertData = {
       allAssociatedUsernames: "bob",
       alertUrl: "https://www.example.com",
-      psk: "wrong_psk",
+      key: "wrong_psk",
       referrer: "https://www.google.com",
       alertType: "reuse",
       suspectedUsername: "testuser",
@@ -34,9 +38,9 @@ describe('API Server Test', () => {
       clientId: "foo"
     };
 
-    const response = await request(alert).post('/alert').send(alertData);
-    expect(response.statusCode).toBe(400);
-    expect(response.body.status).toBe('Incorrect PSK');
+    const response = await axios.post(`${baseURL}/alert`, alertData, {headers});
+    expect(response.status).toBe(400);
+    expect(response.data.status).toBe('Incorrect Key');
   });
 });
